Use eval set to properly expand filenames
Get rid of boolean logic for dry run because it behaves differently than
an if/else.

Check if video file already exists and skip it, so we can leave an
encoded file in place and extract subtitles or audio later.

Use -eq for numerical comparison instead of = which is for string
comparison