Skip symlink if target is a directory
Before this change, the symlink would be deleted, then hard link
creation would fail. Now the symlink will be left in place if it does
not point to a file.

Edit usage to reflect changes; Change var name $file to $link for
clarity; Add explicit -P option to find (just to be safe);