Remove redirect to /dev/null for type cmd
I don't think it was working, and it was definitely breaking things.
After removal, there is no output I can see that needs redirecting. Now
the case statement works as it should, but I don't need to match *is\ a\
function* because I added the `type -t` option which outputs just the
word `function`. This can be matched without asterisks.