Filter out filenames with newlines using find
This protects against symlinks which might have a newline in the filename. It does not protect against a symlink target which might have a newline in the filename. But that's okay, because the target is obtained through realpath, and realpath normalizes paths. The worst that can happen is you get a hard link to the wrong file. If the target doesn't exist, the link won't be created. The check against symlinks makes sure nothing is accidentally deleted due to malformed filenames.