Second safeguard against infinite recursion
Create all folders that will contain hard links BEFORE calling the recursive function, then check for existence of folders before the function is allowed to call itself. This limits recursion even without the first check in place (a find command to see if the symlink exists inside of its target).